From onload to using intervals
The onload mechanism used does not work for Chrome and Edge on
Windows. It works fine for Firefox on Windows, though.

The issue we tried to solve earlier with iframe.onload is to make
sure that iframe.contentWindow is not null before we use it. Now,
we check this manually every 100ms in an interval, removing
dependence on onload.